feat: pixel driller API (#221)
* feat: pixel driller API

The pixel driller API converts many single-band raster files into one gigantic multi-band raster file with each band tagged with its initial file name.
The API will return the value for any given pixel on the `/{x}/{y}` endpoint.

The first time it is run, or anytime the source rasters are updated, the ingestion process will need to be rerun by sending a POST request to the `/reimport` endpoint with `{"code": str}` matching the `ADMIN_CODE` envvar.

* fix: volumes must be a list

* fix: docker-compose setup

* Drop reimport endpoint

* Format with black

* Rewrite pixel_driller with xarray and zarr

* Include metadata in response; change to columnar format

* Test metadata join

* Simplify reshaping

* Expand docs; dict equality

* Check query is within bounds of grids

* Tidy metadata handling

* Small fixes
- make sure `libgdal-dev` is installed during the Docker build.
- update VSCode settings.

* feat: pixel driller frontend (#251)

First pass at a frontend for the pixel-driller API. Clicking anywhere on the map opens a righthand sidebar panel with tables of hazard information for that point.

---------

Co-authored-by: Tom Russell <tom.russell@ouce.ox.ac.uk>
Co-authored-by: Fred Thomas <fred.thomas@eci.ox.ac.uk>
Co-authored-by: Jim O'Donnell <james.odonnell@dtc.ox.ac.uk>