Do not set nodata value for hotspots rasters
We have decided to not set nodata for hotspots rasters. We previously
set non-positive values to nodata so they would be rendered transparent
in the viewer. However, we're now applying a Gaussian kernel to smooth
them and this is leading to small positive values some distance from
assets (and out to sea if the bandwidth is large enough). Rather than
try and identify a threshold below which to set as nodata, we will
retain all the data (do not use nodata) and give the entire layer an
opacity of less than one in the viewer.